+++ propulsion-archiver/Program.fs
@@ -187,7 +187,7 @@ let build (args : Args.Arguments, log) =
     let pipeline =
         use observer = CosmosStoreSource.CreateObserver(log, archiverSink.StartIngester, Seq.collect Handler.selectArchivable)
         let monitored, leases, processorName, startFromTail, maxItems, lagFrequency = args.MonitoringParams()
-        CosmosStoreSource.Run(log, monitored, leases, processorName, observer, startFromTail, ?maxDocuments=maxItems, lagReportFreq=lagFrequency)
+        CosmosStoreSource.Run(log, monitored, leases, processorName, observer, startFromTail, ?maxItems=maxItems, lagReportFreq=lagFrequency)
     archiverSink, pipeline
 
 // A typical app will likely have health checks etc, implying the wireup would be via `endpoints.MapMetrics()` and thus not use this ugly code directly
@@ -202,7 +202,7 @@ let run args = async {
     let sink, pipeline = build (args, log)
     pipeline |> Async.Start
     use _metricsServer : IDisposable = args.PrometheusPort |> Option.map startMetricsServer |> Option.toObj
-    do! sink.AwaitCompletion()
+    do! sink.AwaitWithStopOnCancellation()
 }
 
 [<EntryPoint>]