+++ propulsion-consumer/Program.fs
@@ -74,7 +74,7 @@ let start (args : Args.Arguments) =
 
 let run args = async {
     use consumer = start args
-    return! consumer.AwaitCompletion()
+    return! consumer.AwaitWithStopOnCancellation()
 }
 
 [<EntryPoint>]